Implemented concrete PlotData/PlotSet classes for the data that is to be shown in the plots and its config/metadata.
Reworked the 2x2MagnitudeLogTimePlot so that it's a subclass of the SingleMagnitudeTimePlot so that it uses this super() to do its main rendering in a consistent fashion with the other mag/time plots.

Also used this to introduce the epochs data which will be used to note the spectral epochs on the lighturves. Still need to code the rendering of the epoch data.