Added support for generic "filter" query_param of a DataSource/LightCurve config.
This is used for the XRT rate data to filter out PC data where the count > 1 (as these have high uncertainties).  Should make the XRT curves clearer and easier to fit against.

This has also fixes the issue with the ratio plot where it was showing ratios where there were no visible counts in the rate plots.  This plot is now produced without filtering out PC >1 data and the views are consistent.

Closes #1